% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsFunctions.R
\name{connectToWS}
\alias{connectToWS}
\title{connectToWS}
\usage{
connectToWS(apiID, username, password, url = "")
}
\arguments{
\item{apiID}{character, a character name of an API ("ws_public" or "ws_private")}

\item{username}{login of the user to create the token}

\item{password}{password of the user to create the token}

\item{url}{character, if apiID is private add the url of the chosen API, containing the IP,
the full url with the protocol 'http://www.opensilex.org/openSilexAPI/rest/'}
}
\description{
load name space and connexion parameters of the webservice.
Execute only once at the beginning of the requests.
In the case of a WebService change of address or a renaming of services, please edit this list.
and execute the function.
WS1 - connectToWS(apiID="ws_public","guestphis@supagro.inra.fr","guestphis")
WS2 - connectToWS(apiID="ws_private",username="guest@opensilex.org",password="guest", url = "http://www.opensilex.org/openSilexAPI/rest/")
}
