% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnotations.R
\name{getAnnotations}
\alias{getAnnotations}
\title{getAnnotations}
\usage{
getAnnotations(token, uri = "", creator = "", motivatedBy = "",
  comment = "", target = "", page = NULL, pageSize = NULL,
  verbose = FALSE)
}
\arguments{
\item{token}{character, a token from getToken function}

\item{uri}{character, search by the uri of an annotation (optional)}

\item{creator}{character, search by the uri of th creator of the annotations (optional)}

\item{motivatedBy}{character, search by the motivation to create the annotation ??? (optional)}

\item{comment}{character, search by comment on the annotation (optional)}

\item{target}{character, search by the agronomical object uri that is targeted by an annotation (optional)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}

\item{verbose}{logical FALSE by default, if TRUE display information about the progress}
}
\value{
WSResponse object
}
\description{
retrieves the annotation based on search criterion
}
\details{
You have to execute the getToken() function first to have access to the web
service
}
\examples{
\donttest{
 initializeClientConnection(apiID="ws_private", url = "138.102.159.36:8080/uesAPI/rest/")
 aToken = getToken("guest@phis.fr","guest")
 vars <- getAnnotations(aToken$data,
          uri = "http://www.phenome-fppn.fr/ues/id/annotation/f3ddc6c5-076c-4593-9bce-9efbd4bb7f87")
 vars$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
