% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(variableUri = "", objectUri = "", objectLabel = "",
  provenanceUri = "", provenanceLabel = "", startDate = "",
  endDate = "", page = NULL, pageSize = NULL)
}
\arguments{
\item{variableUri}{character, search by the uri of a variable (NOT optional). You can access the list of variables through \code{\link{getVariables2}} function.}

\item{objectUri}{character, search by object uri}

\item{objectLabel}{character, search by object label}

\item{provenanceUri}{character, search by provenance uri}

\item{provenanceLabel}{character, search by provenance label}

\item{startDate}{character, search from start date (optional)}

\item{endDate}{character, search to end date (optional)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
Retrieves the data from the web service
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
 vars<-getVariables2()$data$uri
 totalCount <- getData(variable = vars[4])$totalCount
 data <- getData(variable = vars[4],pageSize = totalCount)
 data <- getData(variable = vars[5],
                 startDate = "2017-05-01",
                 endDate = "2017-06-01",
                 pageSize = totalCount)
 str(data$data)
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
