% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEvents.R
\name{getEvents}
\alias{getEvents}
\title{getEvents}
\usage{
getEvents(token, uri = "", type = "", concernsUri = "",
  concernsLabel = "", dateRangeStart = "", dateRangeEnd = "",
  page = NULL, pageSize = NULL, verbose = FALSE)
}
\arguments{
\item{token}{character, a token from getToken function}

\item{uri}{character, search by the uri of an event (optional)}

\item{type}{character, search by the type of an event (optional)}

\item{concernsUri}{character, search by the URI concerned by the event (optional)}

\item{concernsLabel}{character, search by the label concerned by the event (optional)}

\item{dateRangeStart}{character, search from the start of range (optional)}

\item{dateRangeEnd}{character, search to the end of range (optional)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}

\item{verbose}{logical, FALSE by default, if TRUE display information about the progress}
}
\value{
WSResponse object
}
\description{
retrieves the events based on search criterion
}
\details{
You have to execute the getToken() function first to have access to the web
service
}
\examples{
\donttest{
initializeClientConnection(apiID="ws_private", url = "www.opensilex.org/openSilexAPI/rest/")
aToken = getToken("guest@opensilex.org","guest")
events <- getEvents(aToken$data,
 type = "http://www.opensilex.org/vocabulary/oeev#Fertilization",
  pageSize=10)
events$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
