% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperiments.R
\name{getExperimentById}
\alias{getExperimentById}
\title{getExperimentById}
\usage{
getExperimentById(experimentURI = "", page = NULL, pageSize = NULL)
}
\arguments{
\item{experimentURI}{URI of the experiment}

\item{page}{displayed page (pagination Plant Breeding API)}

\item{pageSize}{number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
retrieves the informations for one experiment
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
 connectToPHISWS(apiID="ws_1_public", 
                 username = "guestphis@supagro.inra.fr",
                 password = "guestphis")
 Exp1<-getExperimentById(
        experimentURI ="http://www.phenome-fppn.fr/m3p/ARCH2012-01-01")
 Exp1$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
\keyword{internal}
