% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperiments.R
\name{getExperiments}
\alias{getExperiments}
\title{retrieves the experiments from the web service}
\usage{
getExperiments(projectName = "", season = "", sortOrder = "DESC",
  page = NULL, pageSize = NULL)
}
\arguments{
\item{projectName}{project name}

\item{season}{character, a year when the experiment was conducted}

\item{sortOrder}{ordering "ASC" or "DESC"}

\item{page}{displayed page (pagination Plant Breeding API)}

\item{pageSize}{number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
Retrieves the available experiments and/or linked to a project
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
 connectToPHISWS(apiID="ws_1_public", 
                 username = "guestphis@supagro.inra.fr",
                 password = "guestphis")
 getExperiments(page=3)
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
