% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperiments.R
\name{getExperiments2}
\alias{getExperiments2}
\title{getExperiments2}
\usage{
getExperiments2(
  uri = "",
  startDate = "",
  endDate = "",
  field = "",
  campaign = "",
  place = "",
  alias = "",
  keywords = "",
  page = NULL,
  pageSize = NULL
)
}
\arguments{
\item{uri}{character, search by the uri of an experiment (optional)}

\item{startDate}{character, search from the start date of experiment (optional)}

\item{endDate}{character, search to the end date of experiment (optional)}

\item{field}{character, search by the field of an experiment (optional)}

\item{campaign}{character, search by the campaign of an experiment (optional)}

\item{place}{character, search by the place of an experiment (optional)}

\item{alias}{character, search by the alias of an experiment (optional)}

\item{keywords}{character, search by keywords of an experiment (optional)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
retrieves the experiments based on search criterion
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
myexp <- getExperiments2(
              uri = "http://www.opensilex.org/demo/DIA2017-1")
myexp$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
