% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfrastructures.R
\name{getInfrastructures}
\alias{getInfrastructures}
\title{getInfrastructures}
\usage{
getInfrastructures(uri = "", rdfType = "", label = "",
  language = "en", page = NULL, pageSize = NULL)
}
\arguments{
\item{uri}{character, search by the uri of an infrastructure (optional)}

\item{rdfType}{character, search by the rdf type of an infrastructure (optional)}

\item{label}{character, search by the label of an infrastructure (optional)}

\item{language}{character, the language of the response, "en", "fr", etc (optional)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
retrieves the infrastructures based on search criterion
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
test<-getInfrastructures(uri = "https://emphasis.plant-phenotyping.eu")
test$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
