% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws1Calls.R
\name{getLabelViewByExperimentById}
\alias{getLabelViewByExperimentById}
\title{getLabelViewByExperimentById}
\usage{
getLabelViewByExperimentById(experimentURI = "", viewType = "",
  cameraAngle = "", provider = "", page = NULL, pageSize = NULL)
}
\arguments{
\item{experimentURI}{experiment unique identifier}

\item{viewType}{type of view, top or side}

\item{cameraAngle}{numeric, angle of the camera}

\item{provider}{origin of the data}

\item{page}{displayed page (pagination Plant Breeding API)}

\item{pageSize}{number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
Retrieves LabelViews used in a specific experiment
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
# Not run (is an internal function)
 connectToPHISWS(apiID="ws_1_public", 
                 username = "guestphis@supagro.inra.fr",
                 password = "guestphis")
publicLabelView <- getLabelViewByExperimentById(
          experimentURI ="http://www.phenome-fppn.fr/m3p/ARCH2017-11-23")
publicLabelView$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
\keyword{internal}
