% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethods.R
\name{getMethods2}
\alias{getMethods2}
\title{getMethods2}
\usage{
getMethods2(token, uri = "", label = "", page = NULL,
  pageSize = NULL, verbose = FALSE)
}
\arguments{
\item{token}{character, a token from getToken function}

\item{uri}{character, search by the uri of a method (optional)}

\item{label}{character, search by the label of a method (optional)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}

\item{verbose}{logical, FALSE by default, if TRUE display information about the progress}
}
\value{
WSResponse object
}
\description{
retrieves the method based on search criterion
}
\details{
You have to execute the getToken() function first to have access to the web
service
}
\examples{
\donttest{
initializeClientConnection(apiID="ws_private", url = "138.102.159.36:8080/uesAPI/rest/")
aToken = getToken("guest@phis.fr","guest")
vars <- getMethods2(aToken$data,
                   uri = "http://www.phenome-fppn.fr/ues/id/traits/m001")
vars$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
