% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws1Calls.R
\name{getPlantEnvironment}
\alias{getPlantEnvironment}
\title{retrieves the environmental mesures of a plant from the web service}
\usage{
getPlantEnvironment(plantURI = "", variableCategory = "",
  startDate = "", endDate = "", variables = "", facility = "",
  experimentURI = "", page = NULL, pageSize = NULL)
}
\arguments{
\item{plantURI}{character, plant URI}

\item{variableCategory}{character, a category of variables}

\item{startDate}{character, data > startDate (Format: YYYY-MM-DD or YYYY-MM-DD HH:MM:SS )}

\item{endDate}{character, data < startDate (Format: YYYY-MM-DD or YYYY-MM-DD HH:MM:SS )}

\item{variables}{character, list of variables for the request (Ex: "wind speed_weather station_meter per second")}

\item{facility}{character, place of the experiment (Ex: "http://www.phenome-fppn.fr/m3p/ec3")}

\item{experimentURI}{character, URI of the experiment}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
Retrieves environmental mesures of a plant or by dates
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
# not run (is an internal function!!!)
connectToPHISWS(apiID="ws_1_public", 
                 username = "guestphis@supagro.inra.fr",
                 password = "guestphis")
myplant<-getPlantEnvironment(
       plantURI="http://www.phenome-fppn.fr/m3p/arch/2011/c11005809",
       experimentURI = "http://www.phenome-fppn.fr/m3p/ARCH2017-11-23")
myplant$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
\keyword{internal}
