% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws1Calls.R
\name{getPlants}
\alias{getPlants}
\title{retrieves the context of plant linked to an experiment from the web service}
\usage{
getPlants(plantAlias = "", experimentURI = "", germplasmURI = "",
  page = NULL, pageSize = NULL)
}
\arguments{
\item{plantAlias}{character, an alias of plant}

\item{experimentURI}{character, URI of the experiment}

\item{germplasmURI}{character, filter by genotype}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
Retrieves context of plant linked to an experiment
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
 connectToPHISWS(apiID="ws_1_public", 
                 username = "guestphis@supagro.inra.fr",
                 password = "guestphis")
 plantes<-getPlants(experimentURI ="http://www.phenome-fppn.fr/m3p/ARCH2017-03-30")
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
