% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws1Calls.R
\name{getPlantsContextByID}
\alias{getPlantsContextByID}
\title{getPlantsContextByID}
\usage{
getPlantsContextByID(plantURI = "", experimentURI = "", page = NULL,
  pageSize = NULL)
}
\arguments{
\item{plantURI}{character, a URI identifier of a plant}

\item{experimentURI}{URI of the experiment}

\item{page}{displayed page (pagination Plant Breeding API)}

\item{pageSize}{number of elements by page (pagination Plant Breeding API)}

\item{token}{a token}
}
\value{
WSResponse object
}
\description{
Retrieves the contect of a plant
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
# not run (is an internal function!!!)
# aToken<-getToken("guestphis@supagro.inra.fr","guestphis")$data
# test<-getPlantsContextByID(aToken,plantURI="http://www.phenome-fppn.fr/m3p/arch/2011/c11005809",
#       ,experimentURI ="http://www.phenome-fppn.fr/m3p/ARCH2012-01-01")
# test$data
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
\keyword{internal}
