% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws1Calls.R
\name{getProjects}
\alias{getProjects}
\title{getProjects retrieves the list of projects from the web service}
\usage{
getProjects(projectName = "", page = NULL, pageSize = NULL)
}
\arguments{
\item{projectName}{character, Name of the project to search}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
Retrieves the list of projects in the WS
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
 connectToPHISWS(apiID="ws_1_public", 
                 username = "guestphis@supagro.inra.fr",
                 password = "guestphis")
 getProjects()
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
