% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResponseFromWS.R
\name{getResponseFromWS}
\alias{getResponseFromWS}
\title{getResponseFromWS retrieves the data of a service from the WS1}
\usage{
getResponseFromWS(resource, paramPath = NULL, attributes,
  type = "application/json")
}
\arguments{
\item{responseObject}{object HTTP httr}
}
\description{
Create an URL to call the WS and retrun a formatted response of WSResponse class.
}
\keyword{internal}
