% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResponseFromWS.R
\name{getResponseFromWS1}
\alias{getResponseFromWS1}
\title{getResponseFromWS1 retrieves the data of a service from the WS1}
\usage{
getResponseFromWS1(resource, paramPath = NULL, attributes,
  type = "application/json")
}
\arguments{
\item{resource}{character, the name of the service to call}

\item{paramPath}{character, the extension of the service to call, default to NULL}

\item{attributes}{character, the list of attributes to give to the GET request}

\item{type}{character, the type of the output, default to application/json}
}
\value{
responseObject object HTTP httr
}
\description{
Create an URL to call the WS and retrun a formatted response of WSResponse class.
}
\keyword{internal}
