% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResponseFromWS.R
\name{getResponseFromWS2}
\alias{getResponseFromWS2}
\title{getResponseFromWS2 retrieves the data of a service from the WS2}
\usage{
getResponseFromWS2(ressource, paramPath = NULL, attributes,
  type = "application/json", verbose = FALSE)
}
\arguments{
\item{ressource}{character, the name of the service to call}

\item{paramPath}{character, the extension of the service to call, default to NULL}

\item{attributes}{character, The list of attributes to give to the GET request}

\item{type}{character, The type of the output, default to application/json}

\item{verbose}{logical FALSE by default, if TRUE display information about the progress}
}
\description{
Create an URL to call the WS and retrun a formatted response of WSResponse class.
}
\keyword{internal}
