% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScientificObjects.R
\name{getScientificObjects}
\alias{getScientificObjects}
\title{getScientificObjects}
\usage{
getScientificObjects(
  uri = "",
  experiment = "",
  alias = "",
  rdfType = "",
  page = NULL,
  pageSize = NULL
)
}
\arguments{
\item{uri}{character, search by the uri of an annotation (optional)}

\item{experiment}{character, search by the uri of an experiment (optional)}

\item{alias}{character, search by an alias (optional)}

\item{rdfType}{character, search by the rdfType (optional)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
retrieves the scientific objects based on search criterion
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
 test<-getScientificObjects(uri="http://www.opensilex.org/demo/2018/o18000076")
 test$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
