% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSensors.R
\name{getSensorTypes}
\alias{getSensorTypes}
\title{getSensorTypes}
\usage{
getSensorTypes()
}
\value{
WSResponse object
}
\description{
Retrieves the sensors types
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
sensors <- getSensorTypes()
sensors$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
