% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsFunctions.R
\name{getTokenResponseWS}
\alias{getTokenResponseWS}
\title{getTokenResponseWS}
\usage{
getTokenResponseWS(resource, paramPath = NULL, attributes,
  type = "application/json", verbose = FALSE)
}
\arguments{
\item{verbose}{logical FALSE by default, if TRUE display information about the progress}

\item{responseObject}{an object HTTP httr}
}
\description{
Create a token to call the webservice for authentication and
returns a formatted response of WSResponse class.
}
\keyword{internal}
