% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUnits.R
\name{getUnits}
\alias{getUnits}
\title{getUnits}
\usage{
getUnits(token, uri = "", label = "", page = NULL, pageSize = NULL,
  verbose = FALSE)
}
\arguments{
\item{token}{character, a token from getToken function}

\item{uri}{character, search by the uri of a unit (optional)}

\item{label}{character, search by the label of a unit (optional)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}

\item{verbose}{logical, FALSE by default, if TRUE display information about the progress}
}
\value{
WSResponse object
}
\description{
Retrieves the unit based on search criterion
}
\details{
You have to execute the getToken() function first to have access to the web
service
}
\examples{
\donttest{
initializeClientConnection(apiID="ws_private", url = "138.102.159.36:8080/uesAPI/rest/")
aToken = getToken("guest@phis.fr","guest")
vars <- getUnits(aToken$data,
                   uri = "http://www.phenome-fppn.fr/ues/id/units/u001")
vars$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
