% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsFunctions.R
\name{getUserInformations}
\alias{getUserInformations}
\title{getConfigInformations}
\usage{
getUserInformations()
}
\value{
a dataframe with informations from config environment
}
\description{
show useful informations from config environment
}
