% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariables.R
\name{getVariables2}
\alias{getVariables2}
\title{getVariables2}
\usage{
getVariables2(token, uri = "", label = "", trait = "", method = "",
  unit = "", pageSize = NULL, page = NULL, verbose = FALSE)
}
\arguments{
\item{token}{character, a token from getToken function}

\item{uri}{character, search by the uri of an experiment (optional)}

\item{label}{character, search by label (optional)}

\item{trait}{character, search by trait uri (optional)}

\item{method}{character, search by method uri (optional)}

\item{unit}{character, search variables by unit uri (optional)}

\item{pageSize}{character, number of elements by page (pagination Plant Breeding API)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{verbose}{logical, FALSE by default, if TRUE display information about the progress}
}
\value{
WSResponse object
}
\description{
Retrieves the variable descriptions, trait, method and unit covered by the variable
}
\details{
You have to execute the getToken() function first to have access to the web
service
}
\examples{
\donttest{
 initializeClientConnection(apiID="ws_private", url = "138.102.159.36:8080/uesAPI/rest/")
 aToken <- getToken("guest@phis.fr","guest")
 vars <- getVariables2(aToken$data,
          uri = "http://www.phenome-fppn.fr/ues/id/variables/v001")
 vars$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
