% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariables.R
\name{getVariablesByCategory}
\alias{getVariablesByCategory}
\title{getVariablesByCategory}
\usage{
getVariablesByCategory(category = "", experimentURI = "",
  imageryProvider = "", page = NULL, pageSize = NULL)
}
\arguments{
\item{category}{character, Name of the category to search}

\item{experimentURI}{character, URI of the experiment}

\item{imageryProvider}{character, provider of the images}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
Retrieves the variable by categories (environment or setpoint...)
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
 connectToPHISWS(apiID="ws_public", 
                 username = "guestphis@supagro.inra.fr",
                 password = "guestphis")
 vars<-getVariablesByCategory(category="imagery",
          experimentURI = "http://www.phenome-fppn.fr/m3p/ARCH2012-01-01")
 vars$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
