% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariables.R
\name{getVariablesByExperiment}
\alias{getVariablesByExperiment}
\title{getVariablesByExperiment}
\usage{
getVariablesByExperiment(uri = "", pageSize = NULL, page = NULL)
}
\arguments{
\item{uri}{character, search by the uri of an experiment}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}
}
\value{
a WSResponse object. In the 'data' part of the returned object, a data.frame is
   given, containing:
\describe{
\item{trait informations:}{uri, label, comment, ontologiesReferences, properties}
\item{method informations:}{uri, label, comment, ontologiesReferences, properties}
\item{unit informations:}{uri, label, comment, ontologiesReferences, properties}
\item{and uri, label and comment}{for each variable}
}
}
\description{
Retrieves the variable descriptions, trait, method and unit covered by the variable
      for a given experiment URI
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
 varExp<- getVariablesByExperiment(uri="http://www.opensilex.org/demo/DIA2017-1")
 varExp$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
