% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariables.R
\name{getVariablesDetails}
\alias{getVariablesDetails}
\title{getVariablesDetails}
\usage{
getVariablesDetails(uri = "", label = "", trait = "", method = "",
  unit = "", pageSize = NULL, page = NULL, verbose = FALSE)
}
\arguments{
\item{uri}{character, search by the uri of a variable (optional)}

\item{label}{character, search by label (optional)}

\item{trait}{character, search by trait uri (optional)}

\item{method}{character, search by method uri (optional)}

\item{unit}{character, search variables by unit uri (optional)}

\item{pageSize}{character, number of elements by page (pagination Plant Breeding API)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{verbose}{logical, FALSE by default, if TRUE display information about the progress}

\item{token}{character, a token from \code{\link{getToken}} function}
}
\value{
a WSResponse object. In the 'data' part of the returned object, a data.frame is
   given, containing:
\describe{
\item{trait informations:}{uri, label, comment, ontologiesReferences, properties}
\item{method informations:}{uri, label, comment, ontologiesReferences, properties}
\item{unit informations:}{uri, label, comment, ontologiesReferences, properties}
\item{and uri, label and comment}{for each variable}
}
}
\description{
Retrieves the variable descriptions, trait, method and unit covered by the variable
}
\details{
You have to execute the \code{\link{getToken}} function first to have access to the web
service
}
\examples{
\donttest{
 connectToWS(apiID="ws_private", url = "http://www.opensilex.org/openSilexAPI/rest/")
 aToken = getToken("guestphis@supagro.inra.fr","guestphis")
 vars <- getVariablesDetails(aToken$data, uri = "http://www.opensilex.org/demo/id/variables/v001")
 vars <- getVariablesDetails(aToken$data,label = "Leaf-Area_LAI-Computation_LAI")
 vars$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
