% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVectors.R
\name{getVectors}
\alias{getVectors}
\title{getVectors}
\usage{
getVectors(token, uri = "", rdfType = "", label = "", brand = "",
  serialNumber = "", inServiceDate = "", dateOfPurchase = "",
  personInCharge = "", page = NULL, pageSize = NULL,
  verbose = FALSE)
}
\arguments{
\item{token}{character, a token from getToken function}

\item{uri}{character, search by the uri of an annotation (optional)}

\item{rdfType}{character, search by the rdf type of a sensor (optional)}

\item{label}{character, search by the label of the measure (optional)}

\item{brand}{character, search by the brand of a vector (optional)}

\item{serialNumber}{character, search by serialNumber of a vector (optional)}

\item{inServiceDate}{character, search by the inServiceDate of a vector (optional)}

\item{dateOfPurchase}{character, search by the date of purchase of a vector (optional)}

\item{personInCharge}{character, search by the person in charge of a vector (optional)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}

\item{verbose}{logical, FALSE by default, if TRUE display information about the progress}
}
\value{
WSResponse object
}
\description{
retrieves the vectors based on search criterion
}
\details{
You have to execute the getToken() function first to have access to the web
service
}
\examples{
\donttest{
initializeClientConnection(apiID="ws_private", url = "138.102.159.36:8080/uesAPI/rest/")
aToken <- getToken("guest@phis.fr","guest")
vars <- getVectors(aToken$data, uri = "http://www.phenome-fppn.fr/ues/2018/v1801")
vars$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
