% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVocabulary.R
\name{getVocabulary}
\alias{getVocabulary}
\title{getVocabulary}
\usage{
getVocabulary(page = NULL, pageSize = NULL)
}
\arguments{
\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
retrieves the vocabularies from the instance
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
vocabulary<-getVocabulary()
vocabulary$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
