% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postAnnotations.R
\name{postAnnotations}
\alias{postAnnotations}
\title{postAnnotations}
\usage{
postAnnotations(creator, motivatedBy, bodyValues, targets)
}
\arguments{
\item{creator}{character, give the creator of the annotation}

\item{motivatedBy}{character, give the motivation of the annotation, from w3c ontology oa}

\item{bodyValues}{list, The annotation message}

\item{targets}{list, the scientific object(s) targeted by this annotation}
}
\value{
WSResponse object
}
\description{
send an annotation to the web service
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
  postAnnotations(    
    creator = "http://www.opensilex.org/demo/id/agent/marie_dupond",
    motivatedBy = "http://www.w3.org/ns/oa#commenting",
    bodyValues = list("the object has been observed"),
    targets = list("http://www.phenome-fppn.fr/test/2019/o19000074")
  )
     }
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
