% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postData.R
\name{postData}
\alias{postData}
\title{postData}
\usage{
postData(data)
}
\arguments{
\item{data}{a data.frame containing the following columns :
provenanceUri character, the URI of the provenance of the data. Accessed by from \code{\link{getProvenances}}
objectUri character, give the uri of the scientific object concerned
variableUri character, give the uri of the variable concerned
date character, date of the measurement. Follow ISO 8601 format, see example
value character, the value of the measurement}
}
\value{
Response object
}
\description{
send a data to the web service
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
