% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postDataset.R
\name{postDataset}
\alias{postDataset}
\title{postDataset}
\usage{
postDataset(data)
}
\arguments{
\item{data}{data.frame, a dataset with columns names after the arguments of  \code{\link{postData}} function}
}
\value{
WSResponse object
}
\description{
send a dataset to the web service. This extend the \code{\link{postData}} function to a whole dataset
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
  postDataset(
  
  )
   }
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
