% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postEvents.R
\name{postEvents}
\alias{postEvents}
\title{postEvents}
\usage{
postEvents(rdfType, concernedItemsUris, date, properties, description, creator)
}
\arguments{
\item{rdfType}{character, give the rdfType of the event}

\item{concernedItemsUris}{list, give the scientific object targeted by the event}

\item{date}{date, the date of the event. Format ISO8601, see example}

\item{properties}{list, The event properties}

\item{description}{character, description of this event}

\item{creator}{character, the creator of this event}
}
\value{
WSResponse object
}
\description{
send an event to the web service
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
  postEvents(    
    rdfType = "http://www.opensilex.org/vocabulary/oeev#MoveFrom",
    concernedItemsUris = list(),
    date = "2017-09-08T12:00:00+01:00",
    properties = list(
      rdfType = "http://xmlns.com/foaf/0.1/Agent",
      relation = "http://www.opensilex.org/vocabulary/2018#hasContact",
      value= "http://www.opensilex.org/demo/id/agent/marie_dupond"
    ),
    description = "The pest attack lasted 20 minutes",
    creator = "http://www.opensilex.org/demo/id/agent/marie_dupond"
  )
     }
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
