% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postScientificObjects.R
\name{postScientificObjectsTable}
\alias{postScientificObjectsTable}
\title{postScientificObjectsTable}
\usage{
postScientificObjectsTable(table)
}
\arguments{
\item{table}{data.frame, the table of the scientific objects to enter in the Web Service. The table must have the correcponding columns: rdfType, experiment, isPartOf, geometry, properties. In order to facilitate the parsing of data, you can use this function : (une fonction à écrire)}
}
\value{
WSResponse object
}
\description{
send a scientific object to the web service
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\section{Important}{

The properties list must contain the rdfs:label block ; also known as alias.
See example.
}

\examples{
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
expDesign = data.frame(
    Alias = c(
      "1/DZ_PG_67/ZM4394/WW/1/DIA2017-05-19",
      "2/DZ_PG_30/ZM4361/WW/1/DIA2017-05-19",
      "3/DZ_PG_49/ZM4389/WW/1/DIA2017-05-19"),
     Type = c("Plot", "Plot", "Plot"),
     Experiment = c(
     "http://www.opensilex.org/demo/DMO2000-1", 
     "http://www.opensilex.org/demo/DMO2000-1", 
     "http://www.opensilex.org/demo/DMO2000-1"),
     Geometry = c("Polygon ((3.973359402 43.61321499, 3.973389651 43.61319639, 3.973437651 43.6132376, 3.973407402 43.6132562, 3.973359402 43.61321499))", 
     "Polygon ((3.973389651 43.61319639, 3.9734199 43.61317779, 3.973467901 43.61321899, 3.973437651 43.6132376, 3.973389651 43.61319639))", 
     "Polygon ((3.9734199 43.61317779, 3.973450149 43.61315919, 3.97349815 43.61320039, 3.973467901 43.61321899, 3.9734199 43.61317779))"),
     Species = c("Maize", "Maize", "Maize"),
     Variety = c("iPG310", "iPG152", "iPG228")
 )
insert = data.frame(
  rdfType = expDesign$Type,
  experiment = expDesign$Experiment,
  geometry = expDesign$Geometry,
  properties = expDesign$ExperimentModalities
)
   }
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
