% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postSensors.R
\name{postSensor}
\alias{postSensor}
\title{postSensor}
\usage{
postSensor(rdfType, label, brand, model, serialNumber, inServiceDate,
  dateOfPurchase, dateOfLastCalibration, personInCharge)
}
\arguments{
\item{rdfType}{character, the rdfType of the sensor ex: http://www.opensilex.org/vocabulary/oeso#Spectrometer}

\item{label}{character, give the label of this sensor}

\item{brand}{character, the brand of the sensor}

\item{model}{character, the model of the sensor}

\item{serialNumber}{character, the serial number of the sensor}

\item{inServiceDate}{character, date which the sensor has been in service. Format "AAAA-MM-DD"}

\item{dateOfPurchase}{character, date which the sensor has been purchased. Format "AAAA-MM-DD"}

\item{dateOfLastCalibration}{character, date which the sensor has been last calibrated. Format "AAAA-MM-DD"}

\item{personInCharge}{the mail of the person in charge of this sensor}
}
\value{
WSResponse object
}
\description{
send a sensor to the web service
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
postSensor(rdfType = "http://www.opensilex.org/vocabulary/oeso#Spectrometer",
            label = "Sensor_label",
            brand = "Sensor_brand",model = "Sensor_model",serialNumber = "",
            inServiceDate = "2017-06-15",dateOfPurchase = "2017-06-15",
            dateOfLastCalibration = "2017-06-15",personInCharge = "admin@opensilex.org"
)
   }
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
