% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postData.R
\name{postSensorData}
\alias{postSensorData}
\title{postSensorData}
\usage{
postSensorData(data)
}
\arguments{
\item{data}{a data.frame containing the following columns :
provenanceUri character, the URI of the provenance of the data. Accessed by from \code{\link{getProvenances}}
variableUri character, give the uri of the variable concerned
date character, date of the measurement. Follow ISO 8601 format, see example
value character, the value of the measurement}
}
\value{
Response object
}
\description{
same as postData for sensor
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_private", url =
"http://www.opensilex.org/openSilexAPI/rest/",
username="guest@opensilex.org", password="guest") data = data.frame(
provenanceUri = c(
"http://www.phenome-fppn.fr/test/id/provenance/1569422784579",
"http://www.phenome-fppn.fr/test/id/provenance/1569422784579",
"http://www.phenome-fppn.fr/test/id/provenance/1569422784579",
"http://www.phenome-fppn.fr/test/id/provenance/1569422784579"), variableUri
= c( "http://www.phenome-fppn.fr/test/id/variables/v027",
"http://www.phenome-fppn.fr/test/id/variables/v027",
"http://www.phenome-fppn.fr/test/id/variables/v027",
"http://www.phenome-fppn.fr/test/id/variables/v027"), date = c(
"2017-06-15T10:45:00+0200", "2017-06-15T10:46:00+0200",
"2017-06-15T11:47:00+0200", "2017-06-15T11:48:00+0200"), value = c(1111.3,
1010, 3030, 4040) ) postSensorData( data )

}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
