% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postScientificObjects.R, R/postSensors.R
\name{postSensors}
\alias{postSensors}
\title{postVectors}
\usage{
postSensors(uri, rdfType, label, brand, model, serialNumber, inServiceDate,
  dateOfPurchase, personInCharge)

postSensors(uri, rdfType, label, brand, model, serialNumber, inServiceDate,
  dateOfPurchase, personInCharge)
}
\arguments{
\item{rdfType}{character, the rdfType of the scientific object ex: http://www.opensilex.org/vocabulary/oeso#Plot}

\item{label}{character, give the label of this sensor}

\item{brand}{character, the brand of the sensor}

\item{model}{character, the model of the sensor}

\item{serialNumber}{character, the serial number of the sensor}

\item{inServiceDate}{character, date which the sensor has been in service. Format "AAAA-MM-DD"}

\item{dateOfPurchase}{character, date which the sensor has been purchased. Format "AAAA-MM-DD"}

\item{personInCharge}{the mail of the person in charge of this sensor}

\item{alias}{character, give an alias to the scientific object}

\item{geometry}{character, give the geometry of this scientific object. For example a plot can be : "POLYGON((0 0, 10 0, 10 10, 0 10, 0 0))"}

\item{experiment}{character, uri of the experiment of the scientific object}

\item{isPartOf}{character, a scientific object the scientific object is part of}

\item{year}{character, the year of the scientific object}

\item{properties}{list, a list for example: list(rdfType = "http://xmlns.com/foaf/0.1/Agent", 
relation = "http://www.opensilex.org/vocabulary/2018#hasContact", 
value = "http://www.opensilex.org/demo/id/agent/marie_dupond")}

\item{rdfType}{character, the rdfType of the sensor ex: http://www.opensilex.org/vocabulary/oeso#Spectrometer}
}
\value{
WSResponse object

WSResponse object
}
\description{
send a scientific object to the web service

send a sensor to the web service
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service

You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
  postSensors(
   rdfType = "http://www.opensilex.org/vocabulary/oeso#Spectrometer",
   label = "aligot",
   brand = "Fait maison",
   model = "avec du  cantal",
   serialNumber = "",
   inServiceDate = "2017-06-15",
   dateOfPurchase = "2017-06-15",
   dateOfLastCalibration = "2017-06-15",
   personInCharge = "admin@opensilex.org"
   )
   }
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
  postSensors(
   rdfType = "http://www.opensilex.org/vocabulary/oeso#Spectrometer",
   label = "aligot",
   brand = "Fait maison",
   model = "avec du  cantal",
   serialNumber = "",
   inServiceDate = "2017-06-15",
   dateOfPurchase = "2017-06-15",
   dateOfLastCalibration = "2017-06-15",
   personInCharge = "admin@opensilex.org"
   )
   }
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.

http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
