% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postUnits.R
\name{postUnits}
\alias{postUnits}
\title{postUnits}
\usage{
postUnits(uri, label, comment, ontologiesReferences)
}
\arguments{
\item{uri}{character, give a comment for this unit}

\item{label}{character, give the label of this unit}

\item{comment}{character, give a comment for this unit}

\item{ontologiesReferences}{some metadata concerning the ontology of the unit.  The format of the metadata is a list with every item being a different metadata.  Exemple: list(property = "http://www.w3.org/2004/02/skos/core#closeMatch", object = "http://www.cropontology.org/rdf/CO_715:0000139", seeAlso = "http://www.cropontology.org/ontology/CO_715/")}
}
\value{
WSResponse object
}
\description{
send a unit to the web service
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
  postUnits(
   label = "insertionunit_label",
   comment = "comment my unit",
   metadata = list(property = "http://www.w3.org/2004/02/skos/core#closeMatch",
                   object = "http://www.cropontology.org/rdf/CO_715:0000139",
                   seeAlso = "http://www.cropontology.org/ontology/CO_715/")
                   )
                   }
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
