% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postVariables.R
\name{postVariables}
\alias{postVariables}
\title{postVariables}
\usage{
postVariables(uri, label, comment, ontologiesReferences, trait, method,
  unit)
}
\arguments{
\item{uri}{character, give an for this variable. If this is empty, the uri will me automatically generated.}

\item{label}{character, give the label of this variable}

\item{comment}{character, give a comment for this variable}

\item{ontologiesReferences}{some metadata concerning the ontology of the variable.  The format of the metadata is a list with every item being a different metadata.  Exemple: list(property = "http://www.w3.org/2004/02/skos/core#closeMatch", object = "http://www.cropontology.org/rdf/CO_715:0000139", seeAlso = "http://www.cropontology.org/ontology/CO_715/")}

\item{trait}{the uri of the trait composing the variable}

\item{method}{the uri of the method composing the}

\item{unit}{the uri of the unit composing the}
}
\value{
WSResponse object
}
\description{
send a variable to the web service
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
postVariables(uri="http://www.phenome-fppn.fr/mtpvm/2018/v1802",
              label = "insertionvariable_label",
              comment = "comment my variable",
              ontologiesReferences = list(
                property = "http://www.w3.org/2004/02/skos/core#closeMatch",
                object = "http://www.cropontology.org/rdf/CO_715:0000139",
                seeAlso = "http://www.cropontology.org/ontology/CO_715/"),
              trait = "http://www.phenome-fppn.fr/test/id/traits/t008", 
              method = "http://www.phenome-fppn.fr/test/id/methods/m008", 
              unit = "http://www.phenome-fppn.fr/test/id/units/u008")
  }
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
