% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postVectors.R
\name{postVectors}
\alias{postVectors}
\title{postVectors}
\usage{
postVectors(uri, rdfType, label, brand, serialNumber, inServiceDate,
  dateOfPurchase, personInCharge)
}
\arguments{
\item{uri}{character, give the URI of the variable measured by the vector}

\item{rdfType}{character, the rdfType of the vector ex: http://www.opensilex.org/vocabulary/oeso#UAV}

\item{label}{character, give the label of this vector}

\item{brand}{character, the brand of the vector}

\item{serialNumber}{character, the serial number of the vector}

\item{inServiceDate}{character, date which the vector has been in service. Format "AAAA-MM-DD"}

\item{dateOfPurchase}{character, date which the vector has been purchased. Format "AAAA-MM-DD"}

\item{personInCharge}{the mail of the person in charge of this vector}
}
\value{
WSResponse object
}
\description{
send a vector to the web service
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
postVectors(
  uri="http://www.opensilex.org/demo/2019/v1909",
  rdfType = "http://www.opensilex.org/vocabulary/oeso#UAV",
  label = "Vector_label",
  brand = "Vector_brand",
  serialNumber = "1477",
  inServiceDate = "2017-06-15",
  dateOfPurchase = "2017-06-15",
  personInCharge = "admin@opensilex.org"
)
   }
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
