% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsFunctions.R
\name{setLoginUserInformations}
\alias{setLoginUserInformations}
\title{setLoginUserInformations}
\usage{
setLoginUserInformations(username, password, tokenData)
}
\arguments{
\item{tokenData}{S3 class, saves informations extract from WS getToken response}
}
\description{
Save information in config environment
}
\keyword{internal}
