% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnotations.R
\name{getAnnotations}
\alias{getAnnotations}
\title{getAnnotations}
\usage{
getAnnotations(token, uri = "", creator = "", motivatedBy = "",
  comment = "", target = "", page = NULL, pageSize = NULL,
  verbose = FALSE)
}
\arguments{
\item{token}{character, a token from \code{\link{getToken}} function}

\item{uri}{character, search by the uri of an annotation (optional)}

\item{creator}{character, search by the uri of th creator of the annotations (optional)}

\item{motivatedBy}{character, search by the motivation to create the annotation ??? (optional)}

\item{comment}{character, search by comment on the annotation (optional)}

\item{target}{character, search by the agronomical object uri that is targeted by an annotation (optional)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}

\item{verbose}{logical FALSE by default, if TRUE display information about the progress}
}
\value{
WSResponse object
}
\description{
retrieves the annotation based on search criterion
}
\details{
You have to execute the \code{\link{getToken}} function first to have access to the web
service
}
\examples{
\donttest{
 connect(apiID="ws_private",
  url = "www.opensilex.org/openSilexAPI/rest/")
 aToken = getToken("guest@opensilex.org","guest")
 annotations <- getAnnotations(aToken$data,
  comment = "Ustilago maydis infection" , pageSize=10)
 annotations$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
