% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnotations.R
\name{getAnnotations}
\alias{getAnnotations}
\title{getAnnotations}
\usage{
getAnnotations(
  uri = "",
  creator = "",
  motivatedBy = "",
  comment = "",
  target = "",
  page = NULL,
  pageSize = NULL
)
}
\arguments{
\item{uri}{character, search by the uri of an annotation (optional)}

\item{creator}{character, search by the uri of th creator of the annotations (optional)}

\item{motivatedBy}{character, search by the motivation to create the annotation ??? (optional)}

\item{comment}{character, search by comment on the annotation (optional)}

\item{target}{character, search by the agronomical object uri that is targeted by an annotation (optional)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
retrieves the annotation based on search criterion
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
annotations<-getAnnotations(comment = "Ustilago maydis infection" , 
                           pageSize=10)
str(annotations$data)
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
