% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnvironmentData.R
\name{getEnvironment}
\alias{getEnvironment}
\title{retrieves the environmental mesures of an experiment from the web service}
\usage{
getEnvironment(variableCategory = "", startDate = "", endDate = "",
  variables = "", facility = "", experimentURI = "", page = NULL,
  pageSize = NULL)
}
\arguments{
\item{variableCategory}{character, a category of variables}

\item{startDate}{data > startDate (Format: YYYY-MM-DD or YYYY-MM-DD HH:MM:SS )}

\item{endDate}{data < startDate (Format: YYYY-MM-DD or YYYY-MM-DD HH:MM:SS )}

\item{variables}{list of variables for the request (Ex : "wind speed_weather station_meter per second")}

\item{facility}{place of the experiment (Ex : "http://www.phenome-fppn.fr/m3p/ec3")}

\item{experimentURI}{URI of the experiment}

\item{page}{displayed page (pagination Plant Breeding API)}

\item{pageSize}{number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
Retrieves environmental mesures of an experiment or by dates
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
 connectToPHISWS(apiID="ws_public", 
 username = "guestphis@supagro.inra.fr",
 password = "guestphis")
 getEnvironment(page=3,
                pageSize=100,
                startDate="2017-06-29",
                endDate = "2017-06-16")
 test<-getEnvironment(experimentURI="http://www.phenome-fppn.fr/m3p/ARCH2017-03-30")
 test$data
 getEnvironment( experimentURI ="http://www.phenome-fppn.fr/m3p/ARCH2017-03-30",
                 startDate="2017-06-16",
                 endDate="2017-06-29")
 getEnvironment(experimentURI ="http://www.phenome-fppn.fr/m3p/ARCH2017-03-30",
    facility="http://www.phenome-fppn.fr/m3p/es2",
    variables="wind speed_weather station_meter per second")
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
