% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEvents.R
\name{getEvents}
\alias{getEvents}
\title{getEvents}
\usage{
getEvents(uri = "", type = "", concernsUri = "",
  concernsLabel = "", startDate = "", endDate = "", page = NULL,
  pageSize = NULL)
}
\arguments{
\item{uri}{character, search by the uri of an event (optional)}

\item{type}{character, search by the type of an event (optional)}

\item{concernsUri}{character, search by the URI concerned by the event (optional)}

\item{concernsLabel}{character, search by the label concerned by the event (optional)}

\item{startDate}{character, search from the start of range (optional)}

\item{endDate}{character, search to the end of range (optional)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
retrieves the events based on search criterion
}
\details{
You have to execute the \code{\link{connectToWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToWS(apiID="ws_private", url = "http://www.opensilex.org/openSilexAPI/rest/")
events <- getEvents(
 type = "http://www.opensilex.org/vocabulary/oeev#Fertilization",
  pageSize=10)
events$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
