% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperiments.R
\name{getExperimentById}
\alias{getExperimentById}
\title{getExperimentById}
\usage{
getExperimentById(experimentURI = "", page = NULL, pageSize = NULL)
}
\arguments{
\item{experimentURI}{URI of the experiment}

\item{page}{displayed page (pagination Plant Breeding API)}

\item{pageSize}{number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
retrieves the informations for one experiment
}
\details{
You have to execute the getToken() function first to have access to the web
service
}
\examples{
\donttest{
connectToWS(apiID="ws_public")
 publicExp<-getExperimentById(
        experimentURI ="http://www.phenome-fppn.fr/m3p/ARCH2012-01-01")
 publicExp$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
\keyword{internal}
