% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperiments.R
\name{getExperiments}
\alias{getExperiments}
\title{retrieves the experiments from the web service}
\usage{
getExperiments(projectName = "", season = "", sortOrder = "DESC",
  page = NULL, pageSize = NULL)
}
\arguments{
\item{projectName}{project name}

\item{season}{character, a year when the experiment was conducted}

\item{sortOrder}{ordering "ASC" or "DESC"}

\item{page}{displayed page (pagination Plant Breeding API)}

\item{pageSize}{number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
Retrieves the available experiments and/or linked to a project
}
\details{
You have to execute the \code{\link{connectToOpenSILEXWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToOpenSILEXWS(apiID="ws_public")
 getExperiments(page=3,pageSize=100,startDate="2012-02-21",endDate="2012-03-21")
 getExperiments(projectName = "PHIS_Publi")
 getExperiments(sortOrder = "ASC")
 getExperiments(season = 2012 )
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
