% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperiments.R
\name{getExperiments2}
\alias{getExperiments2}
\title{getExperiments2}
\usage{
getExperiments2(token, uri = "", startDate = "", endDate = "",
  field = "", campaign = "", place = "", alias = "",
  keywords = "", page = NULL, pageSize = NULL, verbose = FALSE)
}
\arguments{
\item{token}{character, a token from getToken function}

\item{uri}{character, search by the uri of an experiment (optional)}

\item{startDate}{character, search from the start date of experiment (optional)}

\item{endDate}{character, search to the end date of experiment (optional)}

\item{field}{character, search by the field of an experiment (optional)}

\item{campaign}{character, search by the campaign of an experiment (optional)}

\item{place}{character, search by the place of an experiment (optional)}

\item{alias}{character, search by the alias of an experiment (optional)}

\item{keywords}{character, search by keywords of an experiment (optional)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}

\item{verbose}{logical, FALSE by default, if TRUE display information about the progress}
}
\value{
WSResponse object
}
\description{
retrieves the experiments based on search criterion
}
\details{
You have to execute the getToken() function first to have access to the web
service
}
\examples{
\donttest{
initializeClientConnection(apiID="ws_private", url = "www.opensilex.org/openSilexAPI/rest/")
aToken = getToken("guest@opensilex.org","guest")
myexp <- getExperiments2(aToken$data,
                   uri = "http://www.opensilex.org/demo/DIA2017-1")
myexp$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
