% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfrastructures.R
\name{getInfrastructures}
\alias{getInfrastructures}
\title{getInfrastructures}
\usage{
getInfrastructures(token, uri = "", rdfType = "", label = "",
  language = "en", page = NULL, pageSize = NULL, verbose = FALSE)
}
\arguments{
\item{token}{character, a token from getToken function}

\item{uri}{character, search by the uri of an infrastructure (optional)}

\item{rdfType}{character, search by the rdf type of an infrastructure (optional)}

\item{label}{character, search by the label of an infrastructure (optional)}

\item{language}{character, the language of the response, "en", "fr", etc (optional)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}

\item{verbose}{logical, FALSE by default, if TRUE display information about the progress}
}
\value{
WSResponse object
}
\description{
retrieves the infrastructures based on search criterion
}
\details{
You have to execute the getToken() function first to have access to the web
service
}
\examples{
\donttest{
initializeClientConnection(apiID="ws_private", url = "www.opensilex.org/openSilexAPI/rest/")
aToken = getToken("guest@opensilex.org","guest")
infrastructures <- getInfrastructures(aToken$data,
                   uri = "https://emphasis.plant-phenotyping.eu")
infrastructures$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
