% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethods.R
\name{getMethods2}
\alias{getMethods2}
\title{getMethods2}
\usage{
getMethods2(uri = "", label = "", page = NULL, pageSize = NULL)
}
\arguments{
\item{uri}{character, search by the uri of a method (optional)}

\item{label}{character, search by the label of a method (optional)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
retrieves the method based on search criterion
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_private", url = "http://www.opensilex.org/openSilexAPI/rest/")
methods <- getMethods2(uri = "http://www.opensilex.org/demo/id/methods/m010")
methods$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
