% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethods.R
\name{getMethods2}
\alias{getMethods2}
\title{getMethods2}
\usage{
getMethods2(token, uri = "", label = "", page = NULL,
  pageSize = NULL, verbose = FALSE)
}
\arguments{
\item{token}{character, a token from \code{\link{getToken}} function}

\item{uri}{character, search by the uri of a method (optional)}

\item{label}{character, search by the label of a method (optional)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}

\item{verbose}{logical, FALSE by default, if TRUE display information about the progress}
}
\value{
WSResponse object
}
\description{
retrieves the method based on search criterion
}
\details{
You have to execute the \code{\link{getToken}} function first to have access to the web
service
}
\examples{
\donttest{
initializeClientConnection(apiID="ws_private", url = "www.opensilex.org/openSilexAPI/rest/")
aToken = getToken("guest@opensilex.org","guest")
methods <- getMethods2(aToken$data,
                   uri = "http://www.opensilex.org/demo/id/methods/m010")
methods$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
