% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhenotypeData.R
\name{getPhenotypeData}
\alias{getPhenotypeData}
\title{getPhenotypeData}
\usage{
getPhenotypeData(token, variable = "", experiment = "",
  agroObject = "", startDate = "", endDate = "", sensor = "",
  incertitude = "", page = NULL, pageSize = NULL, verbose = FALSE)
}
\arguments{
\item{token}{character, a token from getToken function}

\item{variable}{character, search by the uri of a variable (optional)}

\item{experiment}{character, search by the uri of an experiment (optional)}

\item{agroObject}{character, search by the uri of an agronomical object (optional)}

\item{startDate}{character, search from start date (optional)}

\item{endDate}{character, search to end date (optional)}

\item{sensor}{character, search by the uri of a sensor (optional)}

\item{incertitude}{character, search by incertitude ???????????????????? (optional)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}

\item{verbose}{logical, FALSE by default, if TRUE display information about the progress}
}
\value{
WSResponse object
}
\description{
Retrieves the phenotype measures data from a variable, an experiment or a sensor
}
\details{
You have to execute the getToken() function first to have access to the web
service
}
\examples{
\donttest{
 initializeClientConnection(apiID="ws_private", url = "www.opensilex.org/openSilexAPI/rest")
 aToken = getToken("guest@opensilex.org","guest")
 phenodata <- getPhenotypeData(aToken$data, variable = "http://www.opensilex.org/demo/id/variables/v010")
 phenodata <- getPhenotypeData(aToken$data, variable = "http://www.opensilex.org/demo/id/variables/v010", startDate="2012-02-21", endDate="2012-02-15 19:20:30")
 phenodata$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
