% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhenotypeData.R
\name{getPhenotypeData}
\alias{getPhenotypeData}
\title{getPhenotypeData}
\usage{
getPhenotypeData(variable = "", experiment = "",
  scientificObject = "", startDate = "", endDate = "", sensor = "",
  incertitude = "", page = NULL, pageSize = NULL, verbose = FALSE)
}
\arguments{
\item{variable}{character, search by the uri of a variable (optional). You can access the list of variables through \code{\link{getVariables2}} function.}

\item{experiment}{character, search by the uri of an experiment (optional). You can access the list of experiments through \code{\link{getExperiments}} function.}

\item{scientificObject}{character, search by the uri of an scientific object (optional). You can access the list of scientific objects through \code{\link{getScientificObjects}} function.}

\item{startDate}{character, search from start date (optional)}

\item{endDate}{character, search to end date (optional)}

\item{sensor}{character, search by the uri of a sensor (optional). you can access the list of sensors through the \code{\link{getSensors}} function.}

\item{incertitude}{character, search by incertitude ???????????????????? (optional)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}

\item{verbose}{logical, FALSE by default, if TRUE display information about the progress}
}
\value{
WSResponse object
}
\description{
Retrieves the phenotype measures data from a variable, an experiment or a sensor
}
\details{
You have to execute the \code{\link{connectToWS}} function first to have access to the web
service
}
\examples{
\donttest{
 connectToWS(apiID="ws_private", url = "http://www.opensilex.org/openSilexAPI/rest/")
 phenodata <- getPhenotypeData(
   variable = "http://www.opensilex.org/demo/id/variables/v001")
 phenodata <- getPhenotypeData(
   variable = "http://www.opensilex.org/demo/id/variables/v001",
   startDate="2017-06-16", endDate="2017-07-16")
 phenodata$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
