% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsCalls.R
\name{getPlants}
\alias{getPlants}
\title{retrieves the context of plant linked to an experiment from the web service}
\usage{
getPlants(plantAlias = "", experimentURI = "", germplasmURI = "",
  page = NULL, pageSize = NULL)
}
\arguments{
\item{plantAlias}{an alias of plant}

\item{experimentURI}{URI of the experiment}

\item{germplasmURI}{filter by genotype}

\item{page}{displayed page (pagination Plant Breeding API)}

\item{pageSize}{number of elements by page (pagination Plant Breeding API)}

\item{token}{a token}
}
\value{
WSResponse object
}
\description{
Retrieves context of plant linked to an experiment
}
\details{
You have to execute the \code{\link{connectToOpenSILEXWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToOpenSILEXWS(apiID="ws_public")
 aToken<-getToken("guestphis@inra.fr","guestphis")$data
 plantes<-getPlants(aToken,experimentURI ="http://www.phenome-fppn.fr/m3p/ARCH2012-01-01")
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
