% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws1Calls.R
\name{getPlantsContextByID}
\alias{getPlantsContextByID}
\title{getPlantsContextByID}
\usage{
getPlantsContextByID(plantURI = "", experimentURI = "", page = NULL,
  pageSize = NULL)
}
\arguments{
\item{plantURI}{character, a URI identifier of a plant}

\item{experimentURI}{character, URI of the experiment}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
Retrieves the contect of a plant
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
# not run (is an internal function!!!)
connectToPHISWS(apiID="ws_1_public", 
                 username = "guestphis@supagro.inra.fr",
                 password = "guestphis")
test<-getPlantsContextByID(
           plantURI="http://www.phenome-fppn.fr/m3p/arch/2011/c11005809",
          experimentURI ="http://www.phenome-fppn.fr/m3p/ARCH2017-11-23")
test$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
\keyword{internal}
