% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjects.R
\name{getProjects}
\alias{getProjects}
\title{getProjects retrieves the list of projects from the web service}
\usage{
getProjects(projectName = "", page = NULL, pageSize = NULL)
}
\arguments{
\item{projectName}{Name of the project to search}

\item{page}{displayed page (pagination Plant Breeding API)}

\item{pageSize}{number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
Retrieves the list of projects in the WS
}
\details{
You have to execute the \code{\link{connectToWS}} function first to have access to the web
service
}
\examples{
\donttest{
 connectToWS(apiID="ws_public")
 getProjects()
 getProjects(page = 1)
 getProjects(page = 3, pageSize = 100)
 getProjects(projectName = "PHIS_Publi")
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
