% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProvenances.R
\name{getProvenances}
\alias{getProvenances}
\title{getProvenance}
\usage{
getProvenances(uri = "", label = "", comment = "",
  jsonValueFilter = "", page = NULL, pageSize = NULL,
  verbose = FALSE)
}
\arguments{
\item{uri}{character, search by the uri of an provenance (optional)}

\item{label}{character, search by the label of an provenance (optional)}

\item{comment}{character, search by the comment in the provenance}

\item{jsonValueFilter}{json object (as character), search by the json value}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}

\item{verbose}{logical, FALSE by default, if TRUE display information about the progress}

\item{token}{character, a token from \code{\link{getToken}} function}
}
\value{
WSResponse object
}
\description{
retrieves the provenance based on search criterion
}
\details{
You have to execute the \code{\link{getToken}} function first to have access to the web
service
}
\examples{
\donttest{
connectToWS(apiID="ws_private", url = "http://www.opensilex.org/openSilexAPI/rest/")
aToken = getToken("guestphis@supagro.inra.fr","guestphis")
provenances <- getProvenances(aToken$data,
 uri = "http://www.opensilex.org/demo/2018/pv181515071552",
  pageSize=10)
provenances <- getProvenances(aToken$data,
 label ="PROV2019-LEAF",
  pageSize=10)
provenances$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
