% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProvenances.R
\name{getProvenances}
\alias{getProvenances}
\title{getProvenance}
\usage{
getProvenances(uri = "", label = "", comment = "",
  jsonValueFilter = NULL, page = NULL, pageSize = NULL)
}
\arguments{
\item{uri}{character, search by the uri of an provenance (optional)}

\item{label}{character, search by the label of an provenance (optional)}

\item{comment}{character, search by the comment in the provenance}

\item{jsonValueFilter}{list, json object (as R list), search by the json value}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
retrieves the provenance based on search criterion
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
provenances <- getProvenances(
   uri = "http://www.opensilex.org/demo/2018/pv181542817185",
   pageSize = 10)
provenances <- getProvenances(
                  label = "MAU17-PG_NDVI_PUBLI",
                  pageSize = 10)
provenances <- getProvenances(
                  label = "PROV2019-LEAF",
                  jsonValueFilter =
                  list(SensingDevice = "http://www.opensilex.org/demo/s001",
                  Vector = "http://www.opensilex.org/demo/v001"),
                  pageSize = 10)
provenances$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
