% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScientificObjects.R
\name{getScientificObjects}
\alias{getScientificObjects}
\title{getScientificObjects}
\usage{
getScientificObjects(token, uri = "", experiment = "", alias = "",
  rdfType = "", page = NULL, pageSize = NULL, verbose = FALSE)
}
\arguments{
\item{token}{character, a token from getToken function}

\item{uri}{character, search by the uri of an annotation (optional)}

\item{experiment}{character, search by the uri of an experiment (optional)}

\item{alias}{character, search by an alias (optional)}

\item{rdfType}{character, search by the rdfType (optional)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}

\item{verbose}{logical, FALSE by default, if TRUE display information about the progress}
}
\value{
WSResponse object
}
\description{
retrieves the scientific objects based on search criterion
}
\details{
You have to execute the getToken() function first to have access to the web
service
}
\examples{
\donttest{
 initializeClientConnection(apiID="ws_private", url = "www.opensilex.org/openSilexAPI/rest")
 aToken = getToken("guest@phis.fr","guest")
 scientificObjects <- getScientificObjects(aToken$data,
                              uri = "http://www.opensilex.org/demo/2018/o18000076")
 scientificObjects$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
