% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpecies.R
\name{getSpecies}
\alias{getSpecies}
\title{getSpecies}
\usage{
getSpecies(token, uri = "", label = "", language = "", page = NULL,
  pageSize = NULL, verbose = FALSE)
}
\arguments{
\item{token}{character, a token from getToken function}

\item{uri}{character, search by the uri of a specie (optional)}

\item{label}{character, search by the label of a specie (optional)}

\item{language}{character, language of the answer, "en", "fr", etc (optional)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}

\item{verbose}{logical, FALSE by default, if TRUE display information about the progress}
}
\value{
WSResponse object
}
\description{
retrieves the specie behind the specie uri
}
\details{
You have to execute the getToken() function first to have access to the web
service
}
\examples{
\donttest{
initializeClientConnection(apiID="ws_private", url = "www.opensilex.org/openSilexAPI/rest/")
aToken = getToken("guest@opensilex.org","guest")
species <- getSpecies(aToken$data)
species$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
