% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpecies.R
\name{getSpecies}
\alias{getSpecies}
\title{getSpecies}
\usage{
getSpecies(uri = "", label = "", language = "", page = NULL,
  pageSize = NULL)
}
\arguments{
\item{uri}{character, search by the uri of a specie (optional)}

\item{label}{character, search by the label of a specie (optional)}

\item{language}{character, language of the answer, "en", "fr", etc (optional)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
retrieves the specie behind the specie uri
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
species <- getSpecies()
species$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
