% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUnits.R
\name{getUnits}
\alias{getUnits}
\title{getUnits}
\usage{
getUnits(uri = "", label = "", page = NULL, pageSize = NULL)
}
\arguments{
\item{uri}{character, search by the uri of a unit (optional)}

\item{label}{character, search by the label of a unit (optional)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
Retrieves the unit based on search criterion
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_private", url = "http://www.opensilex.org/openSilexAPI/rest/")
units <- getUnits(uri="http://www.opensilex.org/demo/id/units/u007")
units$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
