% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUnits.R
\name{getUnits}
\alias{getUnits}
\title{getUnits}
\usage{
getUnits(uri = "", label = "", page = NULL, pageSize = NULL,
  verbose = FALSE)
}
\arguments{
\item{uri}{character, search by the uri of a unit (optional)}

\item{label}{character, search by the label of a unit (optional)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}

\item{verbose}{logical, FALSE by default, if TRUE display information about the progress}

\item{token}{character, a token from \code{\link{getToken}} function}
}
\value{
WSResponse object
}
\description{
Retrieves the unit based on search criterion
}
\details{
You have to execute the \code{\link{getToken}} function first to have access to the web
service
}
\examples{
\donttest{
connectToWS(apiID="ws_private", url = "http://www.opensilex.org/openSilexAPI/rest/")
aToken = getToken("guestphis@supagro.inra.fr","guestphis")
units <- getUnits(aToken$data,
                   uri="http://www.opensilex.org/demo/id/units/u007")
units$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
