% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariables.R
\name{getVariables2}
\alias{getVariables2}
\title{getVariables2 (same as getVariablesDetails function)}
\usage{
getVariables2(
  uri = "",
  label = "",
  trait = "",
  method = "",
  unit = "",
  pageSize = NULL,
  page = NULL
)
}
\arguments{
\item{uri}{character, search by the uri of a variable (optional)}

\item{label}{character, search by label (optional)}

\item{trait}{character, search by trait uri (optional)}

\item{method}{character, search by method uri (optional)}

\item{unit}{character, search variables by unit uri (optional)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}
}
\value{
a WSResponse object. In the 'data' part of the returned object, a data.frame is
   given, containing:
\describe{
\item{trait informations:}{uri, label, comment, ontologiesReferences, properties}
\item{method informations:}{uri, label, comment, ontologiesReferences, properties}
\item{unit informations:}{uri, label, comment, ontologiesReferences, properties}
\item{and uri, label and comment}{for each variable}
}
}
\description{
Retrieves the variable descriptions, trait, method and unit covered by the variable
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
 vars <- getVariables2(uri = "http://www.opensilex.org/demo/id/variables/v001")
 vars <- getVariables2(label = "Leaf-Area_LAI-Computation_LAI")
 vars$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
