% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariables.R
\name{getVariablesByExperiment}
\alias{getVariablesByExperiment}
\title{getVariablesByExperiment}
\usage{
getVariablesByExperiment(uri = "", pageSize = NULL, page = NULL,
  verbose = FALSE)
}
\arguments{
\item{uri}{character, search by the uri of an experiment}

\item{pageSize}{character, number of elements by page (pagination Plant Breeding API)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{verbose}{logical, FALSE by default, if TRUE display information about the progress}
}
\value{
a WSResponse object. In the 'data' part of the returned object, a data.frame is
   given, containing:
\describe{
\item{trait informations:}{uri, label, comment, ontologiesReferences, properties}
\item{method informations:}{uri, label, comment, ontologiesReferences, properties}
\item{unit informations:}{uri, label, comment, ontologiesReferences, properties}
\item{and uri, label and comment}{for each variable}
}
}
\description{
Retrieves the variable descriptions, trait, method and unit covered by the variable
      for a given experiment URI
}
\details{
You have to execute the \code{\link{connectToWS}} function first to have access to the web
service
}
\examples{
\donttest{
 connectToWS(apiID="ws_private","guest@opensilex.org/","guest", url = "http://www.opensilex.org/openSilexAPI/rest/")
 varExp<- getVariablesByExperiment(uri = "http://www.opensilex.org/demo/DIA2017-1")
 varExp$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
