% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVectors.R
\name{getVectors}
\alias{getVectors}
\title{getVectors}
\usage{
getVectors(
  uri = "",
  rdfType = "",
  label = "",
  brand = "",
  serialNumber = "",
  inServiceDate = "",
  dateOfPurchase = "",
  personInCharge = "",
  page = NULL,
  pageSize = NULL
)
}
\arguments{
\item{uri}{character, search by the uri of an annotation (optional)}

\item{rdfType}{character, search by the rdf type of a sensor (optional)}

\item{label}{character, search by the label of the measure (optional)}

\item{brand}{character, search by the brand of a vector (optional)}

\item{serialNumber}{character, search by serialNumber of a vector (optional)}

\item{inServiceDate}{character, search by the inServiceDate of a vector (optional)}

\item{dateOfPurchase}{character, search by the date of purchase of a vector (optional)}

\item{personInCharge}{character, search by the person in charge of a vector (optional)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
retrieves the vectors based on search criterion
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
vectors <- getVectors(uri = "http://www.opensilex.org/demo/2018/v1801")
vectors <- getVectors(serialNumber = "01BD1DD71500001")
vectors$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
