% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws1Calls.R
\name{getWatering}
\alias{getWatering}
\title{retrieves the irrigation data from the web service}
\usage{
getWatering(
  experimentURI = "",
  variablesName = list(),
  provider = "",
  date = "",
  page = NULL,
  pageSize = NULL
)
}
\arguments{
\item{experimentURI}{character, URI of the experiment}

\item{variablesName}{list, variable names of images analysis}

\item{provider}{character, provider of data}

\item{date}{character, data for one day (format: YYYY-MM-DD)}

\item{page}{numeric, displayed page (pagination Plant Breeding API)}

\item{pageSize}{numeric, number of elements by page (pagination Plant Breeding API)}
}
\value{
WSResponse object
}
\description{
Retrieves irrigation data
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
 connectToPHISWS(apiID="ws_1_public", 
                 username = "guestphis@supagro.inra.fr",
                 password = "guestphis")
 mywater<-getWatering(
         experimentURI = "http://www.phenome-fppn.fr/m3p/ARCH2017-11-23",
         variablesName = list("weightBefore"),pageSize=100000)
 head(mywater$data)
}
}
\seealso{
You have to install the opensilexWSClientR before running any 
         request on PHIS web service.

http://docs.brapi.apiary.io/#introduction/url-structure
}
