% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postAnnotations.R
\name{postAnnotations}
\alias{postAnnotations}
\title{postAnnotations}
\usage{
postAnnotations(newAnnotations)
}
\arguments{
\item{newAnnotations}{list, AnnotationDTO list}
}
\value{
WSResponse object
}
\description{
send a list of annotation to the web service
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
               
 newAnnotation <- AnnotationDTO$new(
   creator = "http://www.opensilex.org/demo/id/agent/marie_dupond",
   targets = list("http://www.phenome-fppn.fr/test/2019/o19000074"),
   motivatedBy = "http://www.w3.org/ns/oa#describing",
   bodyValues =  list("the object has been observed")
 )

 response <- postAnnotations(list(newAnnotation))
 response$success
 response$metadata
 }
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
