% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postDataset.R
\name{postDataset}
\alias{postDataset}
\title{postDataset}
\usage{
postDataset(data)
}
\arguments{
\item{data}{data.frame, a dataset with columns names after the arguments of  \code{\link{postData}} function}
}
\value{
WSResponse object
}
\description{
send a dataset to the web service. This extend the \code{\link{postData}} function to a whole dataset
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\examples{
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
data = data.frame(   
  provenanceUri = c(
  "http://www.phenome-fppn.fr/test/id/provenance/1569422784579",
  "http://www.phenome-fppn.fr/test/id/provenance/1569422784579",
  "http://www.phenome-fppn.fr/test/id/provenance/1569422784579",
  "http://www.phenome-fppn.fr/test/id/provenance/1569422784579"),
  objectUri = c(
  "http://www.phenome-fppn.fr/test/2019/o19000076",
  "http://www.phenome-fppn.fr/test/2019/o19000076",
  "http://www.phenome-fppn.fr/test/2019/o19000076",
  "http://www.phenome-fppn.fr/test/2019/o19000076"),
  variableUri = c(
  "http://www.phenome-fppn.fr/test/id/variables/v027",
  "http://www.phenome-fppn.fr/test/id/variables/v027", 
  "http://www.phenome-fppn.fr/test/id/variables/v027",
  "http://www.phenome-fppn.fr/test/id/variables/v027"),
  date = c(
  "2017-06-15T10:45:00+0200",
  "2017-06-15T10:46:00+0200", 
  "2017-06-15T11:47:00+0200",
  "2017-06-15T11:48:00+0200"),
  value = c(1111.3, 1010, 3030, 4040)
)
  postDataset(
    data
  )
   }
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
