% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProvenances.R
\name{postProvenances}
\alias{postProvenances}
\title{postProvenance}
\usage{
postProvenances(label, comment, metadata)
}
\arguments{
\item{label}{character, give the label of this provenance}

\item{comment}{character, give a comment for this provenance}

\item{metadata}{some metadata concerning the provenance. The format of the metadata is a list with every item being a different metadata. Exemple: list(meta1 = "15", meta2 = "owner1")}
}
\value{
WSResponse object
}
\description{
send a provenance to the web service
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
