% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postScientificObjects.R
\name{postScientificObjects}
\alias{postScientificObjects}
\title{postScientificObjects}
\usage{
postScientificObjects(expDesign)
}
\arguments{
\item{expDesign}{a dataframe containing the following columns 
       rdfType character, the rdfType of the scientific object ex: http://www.opensilex.org/vocabulary/oeso#Plot
       Geometry character, give the geometry of this scientific object. For example a plot can be : "POLYGON((0 0, 10 0, 10 10, 0 10, 0 0))"
       Experiment character, uri of the experiment of the scientific object
       parent character, an other scientific object the disignated scientific object is part of
       properties list, a column for the different properties avaliable. To see the different properties use \code{\link{getVocabulary}} 
Note: The object can also be a data.frame with the 3 named column rdfType, relation and value. See example.}
}
\value{
WSResponse object
}
\description{
send a scientific object to the web service
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\section{Important}{

The properties list must contain the rdfs:label block ; also known as Alias.
See example.
}

\examples{
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
expDesign = data.frame(
     Alias = c(
      "1/DZ_PG_67/ZM4394/WW/1/DIA2017-05-19",
      "2/DZ_PG_30/ZM4361/WW/1/DIA2017-05-19",
      "3/DZ_PG_49/ZM4389/WW/1/DIA2017-05-19"),
     Type = c("Plot", "Plot", "Plot"),
     Experiment = c(
     "http://www.opensilex.org/demo/DMO2000-1", 
     "http://www.opensilex.org/demo/DMO2000-1", 
     "http://www.opensilex.org/demo/DMO2000-1"),
     Species = c("Maize", "Maize", "Maize"), 
     Variety = c("iPG310", "iPG152", "iPG228") 
 )
   }
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
