% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postScientificObjects.R
\name{postScientificObjects}
\alias{postScientificObjects}
\title{postScientificObjects}
\usage{
postScientificObjects(
  rdfType,
  experiment,
  geometry = "",
  isPartOf = "",
  properties
)
}
\arguments{
\item{rdfType}{character, the rdfType of the scientific object ex: http://www.opensilex.org/vocabulary/oeso#Plot}

\item{experiment}{character, uri of the experiment of the scientific object}

\item{geometry}{character, give the geometry of this scientific object. For example a plot can be : "POLYGON((0 0, 10 0, 10 10, 0 10, 0 0))"}

\item{isPartOf}{character, a scientific object the scientific object is part of ???}

\item{properties}{list, a list for the properties. 
Note: The object can also be a data.frame with the 3 named column rdfType, relation and value. See example.}
}
\value{
WSResponse object
}
\description{
send a scientific object to the web service
}
\details{
You have to execute the \code{\link{connectToPHISWS}} function first to have access to the web
service
}
\section{Important}{

This list must contain the rdfs:label block ; also known as alias.
See example.
}

\examples{
\donttest{
connectToPHISWS(apiID="ws_private",
               url = "http://www.opensilex.org/openSilexAPI/rest/",
               username="guest@opensilex.org",
               password="guest")
postScientificObjects(
  rdfType = "http://www.opensilex.org/vocabulary/oeso#Plot",
  experiment="http://www.opensilex.org/demo/DIA2017-1",
  properties=list(
    list(
      rdfType = "http://www.opensilex.org/vocabulary/oeso#Species",
      relation = "http://www.opensilex.org/vocabulary/oeso#hasSpecies",
      value = "http://www.phenome-fppn.fr/id/species/triticumaestivum"),
    list(
      rdfType = NA,
      relation ="http://www.w3.org/2000/01/rdf-schema#label",
      value ="objectAlias")
)
)
   }
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure

You have to install the opensilexWSClientR before running any 
         request on PHIS web service.
}
