% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsFunctions.R
\name{setDebugMode}
\alias{setDebugMode}
\title{setDebugMode}
\usage{
setDebugMode(active = TRUE)
}
\arguments{
\item{active}{logical, FALSE by default, if TRUE display information about the progress}
}
\description{
Activate or deactivate debug mode
}
